+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link'
 import { NAV_ITEMS } from './(data)/constants'
-import { ArrowRightIcon } from '@heroicons/react/24/outline'
+import { ArrowRightIcon } from '@heroicons/react/24/solid'
 import PageActions from './(components)/PageActions'
 import BackToTopButton from './(components)/BackToTopButton'
 import Image from 'next/image'